pyproject.toml
[tool.coverage.run]
branch = true
source = ['street_food_api']
omit = [
  '*/migrations/*',
  '*/manage.py',
  '*/asgi.py',
  '*/wsgi.py',
  '*/settings/*',
  '*/__init__.py'
]

[tool.isort]
known_third_party = ["celery", "django", "django_filters", "factory", "freezegun", "phonenumber_field", "pytest", "requests", "rest_framework"]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
known_first_party = ["street_food_api", "locations", "trucks", "conf"]
known_django = "django"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
  | migrations
)/
'''

[tool.poetry]
name = "streed_food_api"
version = "0.1.0"
description = ""
authors = ["Adam Szustak <madasz@onet.eu>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.5"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.8.6"
django-phonenumber-field = "^5.0.0"
pip = "^20.3.3"
install = "^1.3.4"
phonenumbers = "^8.12.16"
Pillow = "^8.1.0"
django-cleanup = "^5.1.0"
django-filter = "^2.4.0"
freezegun = "^1.1.0"
celery = "^5.0.5"
requests = "^2.25.1"
redis = "^3.5.3"
django-redis = "^4.12.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
pre-commit = "^2.9.3"
pytest-pythonpath = "^0.7.3"
factory-boy = "^3.2.0"
pytest = "^6.2.2"
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"
coverage = {extras = ["toml"], version = "^5.4"}
tox = "^3.21.3"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
